moved stateful component in container
